Incorrect progress upload percent in MODERN

+++ pangaea-backend/cores/core.h
@@ -65,10 +65,10 @@ private:
     QList<QByteArray> commandsSended;
     quint16 commandCount;
 
-    quint32 symbolsToSend{0};
-    quint32 symbolsSended{0};
-    quint32 symbolsToRecieve{0};
-    quint32 symbolsRecieved{0};
+    qint64 symbolsToSend{0};
+    qint64 symbolsSended{0};
+    qint64 symbolsToRecieve{0};
+    qint64 symbolsRecieved{0};
 
     void calculateSendVolume();
     void updateProgressBar();
